Make new branch for the old password validation example
