+++ src/router/routes.tsx
@@ -23,25 +23,29 @@ export const PAGES: Routes = {
     },
     FORM_VALIDATION: {
         title: 'Forms and Validation',
-        LIST: {
-            title: 'In a List',
-            route: 'in-a-list',
-        },
         FIXED_LENGTH_PASSCODE: {
             title: 'Fixed Length Passcode',
             route: 'fixed-length-passcode',
+        },
+        LIST: {
+            title: 'In a List',
+            route: 'in-a-list',
         },
         TABLE: {
             title: 'In a Table',
             route: 'in-a-table',
         },
-        SECTIONED_FORM: {
-            title: 'Sectioned Form',
-            route: 'in-a-sectioned-form',
+        PASSWORD: {
+            title: 'Password',
+            route: 'password',
         },
         PHONE_NUMBER_FORMAT: {
             title: 'Phone Number Format',
             route: 'phone-number-format',
+        },
+        SECTIONED_FORM: {
+            title: 'Sectioned Form',
+            route: 'in-a-sectioned-form',
         },
         VERIFY_ON_SUBMIT: {
             title: 'Verify on Submit',
@@ -159,6 +163,7 @@ export type Routes = {
         VERIFY_ON_SUBMIT: RouteMetaData;
         PHONE_NUMBER_FORMAT: RouteMetaData;
         TABLE: RouteMetaData;
+        PASSWORD: RouteMetaData;
     };
     I18N: RouteMetaData;
     LISTS: RouteMetaData & {
